ClipVitB16-v1.0/requirements.txt
torch
torchvision
transformers
matplotlib
pillow